Move common driver setup steps to the base app driver.
